Moved fileCache into separate caching package
